+++ src/Model/Config/Exception/YamlStandardConfigExceptionInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace YamlStandards\Model\Config\Exception;
+
+use Throwable;
+
+interface YamlStandardConfigExceptionInterface extends Throwable
+{
+}