+++ src/Model/Config/Exception/YamlStandardConfigNotFoundException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace YamlStandards\Model\Config\Exception;
+
+class YamlStandardConfigNotFoundException extends \Exception implements YamlStandardConfigExceptionInterface
+{
+}